chore: moved SBBSegmentedButton to beta
